// playwright-e2e/constants/ccd-events/claimant-response-spec-type/claimant-response-spec-type.ts
enum ClaimantResponseSpecType {
  ACCEPT_FULL_ADMIT = 'ACCEPT_FULL_ADMIT',
  REJECT_FULL_DEFENCE = 'REJECT_FULL_DEFENCE',
  ACCEPT_FULL_DEFENCE = 'ACCEPT_FULL_DEFENCE',
  REJECT_PART_ADMIT = 'REJECT_PART_ADMIT',
  ACCEPT_PART_ADMIT = 'ACCEPT_PART_ADMIT',
  REJECT_PART_ADMIT_PAID_CONFIRM_PAID = 'REJECT_PART_ADMIT_PAID_CONFIRM_HAS_PAID',
  ACCEPT_PART_ADMIT_PAID_CONFIRM_HAS_PAID = 'ACCEPT_PART_ADMIT_PAID_CONFIRM_HAS_PAID',
  REJECT_PART_ADMIT_PAID_CONFIRM_NOT_PAID = 'REJECT_PART_ADMIT_PAID_CONFIRM_HAS_NOT_PAID',
}

export default ClaimantResponseSpecType;